pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tuda"
version = "0.1.0"
description = "TUDA hybrid app backend services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "google-api-python-client>=2.150.0",
  "google-auth>=2.35.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
db = ["pysqlcipher3-binary>=1.0.4"]
ml = ["transformers>=4.45.0", "bitsandbytes>=0.44.0", "torch>=2.4.0"]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.7.0",
  "black>=24.10.0",
]

[tool.setuptools]
packages = ["tuda", "tuda.plugins"]

[tool.pytest.ini_options]
addopts = "--cov=tuda --cov-report=term-missing --cov-fail-under=80"
testpaths = ["tests"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
extend-ignore = ["B008"]
